fix: move langauge selection inside account page
