+++ public/i18n/en/translation.json
@@ -830,7 +830,8 @@
     "account_management": {
       "title": "Account management for '{name}'",
       "change_password": "Change password",
-      "change_password_description": "Note: if you change the password for 'root' user, you will change the password for shell access too."
+      "change_password_description": "Note: if you change the password for 'root' user, you will change the password for shell access too.",
+      "language": "language"
     },
     "change_password": {
       "password_rules": "Must be 8 characters or longer, contain at least one lowercase letter, one uppercase letter, a number and a symbol",
