+++ src/views/standalone/UserManagement.vue
@@ -4,7 +4,7 @@
 -->
 
 <script lang="ts" setup>
-import { NeTitle } from '@nethserver/vue-tailwind-lib'
+import { NeButton, NeTitle, savePreference } from '@nethserver/vue-tailwind-lib'
 import { useI18n } from 'vue-i18n'
 import ChangePassword from '@/components/standalone/account/ChangePassword.vue'
 import { useLoginStore } from '@/stores/standalone/standaloneLogin'
@@ -13,6 +13,13 @@ import FormLayout from '@/components/standalone/FormLayout.vue'
 const { t } = useI18n()
 
 const loginStore = useLoginStore()
+
+async function changeLocale(lang: string) {
+  savePreference('locale', lang, loginStore.username)
+
+  // reload page
+  location.reload()
+}
 </script>
 
 <template>
@@ -23,6 +30,10 @@ const loginStore = useLoginStore()
       :description="t('standalone.account_management.change_password_description')"
     >
       <ChangePassword />
+    </FormLayout>
+    <FormLayout :title="t('standalone.account_management.language')">
+      <NeButton size="lg" @click="changeLocale('it')" class="mb-4 mr-4">ITA</NeButton>
+      <NeButton size="lg" @click="changeLocale('en')" class="mb-4">ENG</NeButton>
     </FormLayout>
   </div>
 </template>
